账号管理开发完成

+++ src/main/java/com/example/demo/biz/AccountBiz.java
@@ -8,10 +8,10 @@ import com.example.demo.base.response.RpaResponse;
 import com.example.demo.constant.CommonConstant;
 import com.example.demo.constant.RedisPrefix;
 import com.example.demo.dao.AccountMapper;
+import com.example.demo.dao.AccountRecordMapper;
 import com.example.demo.dao.RobotMapper;
-import com.example.demo.entity.Account;
-import com.example.demo.entity.Robot;
-import com.example.demo.entity.RobotVo;
+import com.example.demo.entity.*;
+import com.example.demo.util.JsonUtil;
 import com.example.demo.util.StringUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.redisson.api.RQueue;
@@ -20,9 +20,9 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import javax.security.auth.login.Configuration;
 import java.util.Date;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Slf4j
 @Service
@@ -36,6 +36,9 @@ public class AccountBiz {
 
     @Resource
     private RedissonClient redisson;
+
+    @Resource
+    private AccountRecordMapper accountRecordMapper;
 
     private static final Snowflake snowflake = IdUtil.getSnowflake(1, 1);
 
@@ -61,37 +64,76 @@ public class AccountBiz {
      * @param account
      * @return
      */
-    public Account getAccountByWebSite(Account account) {
-        // 根据网站名 在redis 队列里面获取一个账号
-        RQueue<String> accountQ = redisson.getQueue(RedisPrefix.ACCOUNT + account.getWebsite());
-        String accountJson = accountQ.poll();
-        // 不为空的话更新状态
-        if (StringUtils.isNotEmpty(accountJson)) {
-            Account acc = JSONObject.parseObject(accountJson, Account.class);
-            acc.setRobotId(account.getRobotId());
-            acc.setTakeTime(new Date());
-            acc.setUpdateTime(new Date());
-            accountMapper.updateStatusById(acc);
-            return acc;
+    public TakeAccountDto getAccountByWebSite(TakeAccountDto account) {
+
+        // 根据机器人id 判断 是否已经领用过账号
+        List<AccountRecord> recordList = accountRecordMapper.selectRobotUsed(account.getRobotId(), CommonConstant.USED);
+
+        // 没有领用记录
+        if (recordList.size() == 0) {
+            // 根据网站名 在redis 队列里面获取一个账号
+            RQueue<String> accountQ = redisson.getQueue(RedisPrefix.ACCOUNT + account.getWebsite());
+            String accountJson = accountQ.poll();
+            // 不为空的话更新状态
+            if (StringUtils.isNotEmpty(accountJson)) {
+                Account acc = JSONObject.parseObject(accountJson, Account.class);
+                // 新增账号记录
+                AccountRecord accountRecord = new AccountRecord();
+                Long recordId = snowflake.nextId();
+                accountRecord.setId(recordId);
+                accountRecord.setRobotId(account.getRobotId());
+                accountRecord.setTaskId(account.getTaskId());
+                accountRecord.setAccountId(acc.getId());
+                accountRecord.setUseTime(new Date());
+                accountRecord.setStatus(CommonConstant.USED);
+                accountRecordMapper.insertSelective(accountRecord);
+                // 添加返回给机器人的信息
+                account.setAccount(acc.getAccount());
+                account.setPwd(acc.getPwd());
+                account.setRecordId(recordId);
+                return account;
+            } else {
+                throw new RpaException(RpaResponse.getErrorResult("30069"));
+            }
         } else {
-            throw new RpaException(RpaResponse.getErrorResult("30069"));
+            AccountRecord record = recordList.get(0);
+            // 有领用记录,根据记录的账号返回给机器人
+            Account odlAcc = accountMapper.selectById(record.getAccountId());
+            account.setAccount(odlAcc.getAccount());
+            account.setPwd(odlAcc.getPwd());
+            account.setRecordId(record.getId());
+            return account;
         }
     }
 
     /**
      * 归还账号
      *
-     * @param account
+     * @param takeDto
      */
-    public void restoreAccount(Account account) {
-        Account acc = accountMapper.selectByPrimaryKey(account.getId());
-        acc.setRobotId(null);
-        acc.setTakeTime(null);
-        acc.setUpdateTime(new Date());
-        accountMapper.updateStatusById(acc);
+    public void restoreAccount(TakeAccountDto takeDto) {
+        Account acc = accountMapper.selectByPrimaryKey(takeDto.getAccountId());
+        if (acc == null) {
+            // 归还了无效的账号
+            throw new RpaException(RpaResponse.getErrorResult("30070"));
+        }
         // 添加到redis
-        RQueue<Object> accountQ = redisson.getQueue(RedisPrefix.ACCOUNT + acc.getWebsite());
-        accountQ.add(JSONObject.toJSONString(account));
+        RQueue<String> accountQ = redisson.getQueue(RedisPrefix.ACCOUNT + acc.getWebsite());
+        // 查询redis 里面是否已经有该账号，避免重复归还
+        boolean exist = accountQ.stream().anyMatch(ac -> {
+            Account account = JsonUtil.parseObject(ac, Account.class);
+            return account.getId().equals(takeDto.getAccountId());
+        });
+        // 已经存在
+        if (!exist) {
+            // 更新账号记录 、 归还账号
+            AccountRecord accountRecord = new AccountRecord();
+            accountRecord.setStatus(CommonConstant.UN_USED);
+            accountRecord.setBackTime(new Date());
+            accountRecord.setId(takeDto.getRecordId());
+            accountRecordMapper.updateStatusById(accountRecord);
+            accountQ.add(JSONObject.toJSONString(acc));
+        }
     }
 
     /**
@@ -101,42 +143,27 @@ public class AccountBiz {
      */
     public void deleteAccount(Account account) {
         Account acc = accountMapper.selectByPrimaryKey(account.getId());
-        RQueue<String> accountQ = redisson.getQueue(RedisPrefix.ACCOUNT + acc.getWebsite());
-        accountMapper.deleteAccount(account);
-        // 根据id 删除队列里面的账号
-        accountQ.removeIf(s -> s.contains(acc.getId().toString()));
-    }
-
-    /**
-     * 查看数据库中领用账号信息是否正确
-     *
-     */
-    @Scheduled(cron = "0 0/5 * * * ?")
-    public void fixAccount() {
-        log.info("查询账号领用情况是否无误");
-        List<Account> accounts = accountMapper.selectUsedAccount();
-        Robot robot = new Robot();
-        robot.setStatus(CommonConstant.ROBOT_STATUS_BUSY);
-        List<RobotVo> robotVoList = robotMapper.selectRobotList(robot);
-        // 查找被领用的账号，但是机器人没有使用的情况
-        List<Account> accountList = accounts.stream().filter(acc -> {
-            return robotVoList.stream().anyMatch(bot -> {
-                return bot.getId().equals(acc.getRobotId());
-            });
-        }).collect(Collectors.toList());
-
-        if (accountList.size() > 0) {
-            // 批量更新账号
-            accountMapper.updateStatusBatch(accountList);
-            // 账号重新添加到redis 队列
-            for (Account acc : accountList) {
-                RQueue<String> accountQ = redisson.getQueue(RedisPrefix.ACCOUNT + acc.getWebsite());
-                accountQ.add(JSONObject.toJSONString(acc));
-            }
+        if (acc != null) {
+            RQueue<String> accountQ = redisson.getQueue(RedisPrefix.ACCOUNT + acc.getWebsite());
+            accountMapper.deleteAccount(account);
+            // 根据id 删除队列里面的账号
+            accountQ.removeIf(s -> s.contains(acc.getId().toString()));
         }
-
-
     }
 
+    public void timeOutBackAccount(Long robotId) {
+        // 判断是否有正在使用的账号但是没有归还
+        List<AccountRecord> accountRecordList = accountRecordMapper.selectRobotUsed(robotId, CommonConstant.USED);
+        if (accountRecordList.size() > 0) {
+            AccountRecord accountRecord = accountRecordList.get(0);
+            Account account = accountMapper.selectById(accountRecord.getAccountId());
+            // 添加到redis
+            RQueue<String> accountQ = redisson.getQueue(RedisPrefix.ACCOUNT + account.getWebsite());
+            accountQ.add(JSONObject.toJSONString(account));
+            // 手动归还
+            accountRecord.setStatus(CommonConstant.UN_USED);
+            accountRecordMapper.updateStatusById(accountRecord);
+        }
+    }
 
 }
